CLUB DATA
𝗔𝗜𝗠:
In this Project with Microsoft Excel, I explore the dataset to find possible visualizations in other to help the stakeholders to arrive at a good decision making.

𝗪𝗵𝗮𝘁 𝗶𝘀 𝘁𝗵𝗲 𝗔𝗴𝗲𝗻𝗱𝗮 𝗼𝗳 𝘁𝗵𝗲 𝗽𝗿𝗼𝗷𝗲𝗰𝘁?
The project’s Agenda involves Analyzing the data using Microsoft Excel. I performed the following activties
Stage 1: Standardizing the Dataset – This stage tests the ability to process data between different types of variables.

Stage 2: Conversions – This stage tests the ability to convert formulas (i.e., centimeters to inches) and use the VLOOKUP function to populate them.

Stage 3: Pivot Table 1 – This stage tests the ability to create a pivot table.

Stage 4: Pivot Table 2 – This stage tests the ability to take an existing pivot table and change its values, creating new value tabs.

Stage 5: Sorting and Pivotchart – This stage tests the ability to sort data within a pivot table or pivotchart.

Stage 6: Print Setup and Output – This stage tests the ability to create a header using Page Setup, as well as print the document to a PDF file.